separate out nested classes and interfaces
