+++ src/System.Windows.Forms/src/System/Windows/Forms/PropertyGridInternal/PropertyGridView.cs
@@ -7,7 +7,6 @@ using System.ComponentModel;
 using System.ComponentModel.Design;
 using System.Diagnostics;
 using System.Drawing;
-using System.Drawing.Design;
 using System.Drawing.Drawing2D;
 using System.Globalization;
 using System.Runtime.InteropServices;
@@ -18,7 +17,7 @@ using static Interop;
 
 namespace System.Windows.Forms.PropertyGridInternal
 {
-    internal class PropertyGridView :
+    internal partial class PropertyGridView :
     Control,
     IWin32Window,
     IWindowsFormsEditorService,
@@ -6316,2645 +6315,6 @@ namespace System.Windows.Forms.PropertyGridInternal
                 }
             }
             return;
-        }
-
-        private class DropDownHolder : Form, IMouseHookClient
-        {
-            private Control currentControl = null; // the control that is hosted in the holder
-            private readonly PropertyGridView gridView;              // the owner gridview
-            private readonly MouseHook mouseHook;             // we use this to hook mouse downs, etc. to know when to close the dropdown.
-
-            private LinkLabel createNewLink = null;
-
-            // all the resizing goo...
-            //
-            private bool                            resizable       = true;  // true if we're showing the resize widget.
-            private bool                            resizing        = false; // true if we're in the middle of a resize operation.
-            private  bool                           resizeUp        = false; // true if the dropdown is above the grid row, which means the resize widget is at the top.
-            private Point                           dragStart       = Point.Empty;     // the point at which the drag started to compute the delta
-            private Rectangle                       dragBaseRect    = Rectangle.Empty; // the original bounds of our control.
-            private int                             currentMoveType = MoveTypeNone;    // what type of move are we processing? left, bottom, or both?
-        
-            private int             ResizeBarSize;    // the thickness of the resize bar
-            private int             ResizeBorderSize; // the thickness of the 2-way resize area along the outer edge of the resize bar
-            private int             ResizeGripSize;   // the size of the 4-way resize grip at outermost corner of the resize bar
-            private Size            MinDropDownSize;  // the minimum size for the control.
-            private Bitmap                          sizeGripGlyph;    // our cached size grip glyph.  Control paint only does right bottom glyphs, so we cache a mirrored one.  See GetSizeGripGlyph
-
-            private const int                       DropDownHolderBorder = 1;
-            private const int                       MoveTypeNone    = 0x0;
-            private const int                       MoveTypeBottom	= 0x1;
-            private const int                       MoveTypeLeft	= 0x2;
-            private const int                       MoveTypeTop  	= 0x4;
-
-            
-            
-            // DropDownHolder's caption is not visible,  So we don't have to localize its text.
-            [SuppressMessage("Microsoft.Globalization", "CA1303:DoNotPassLiteralsAsLocalizedParameters")]
-            internal DropDownHolder(PropertyGridView psheet)
-            : base() {
-                MinDropDownSize = new Size(SystemInformation.VerticalScrollBarWidth * 4, SystemInformation.HorizontalScrollBarHeight * 4);
-                ResizeGripSize = SystemInformation.HorizontalScrollBarHeight;
-                ResizeBarSize = ResizeGripSize + 1;
-                ResizeBorderSize = ResizeBarSize / 2;
-
-                this.ShowInTaskbar = false;
-                this.ControlBox = false;
-                this.MinimizeBox = false;
-                this.MaximizeBox = false;
-                this.Text = "";
-                this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.None;
-                this.AutoScaleMode = AutoScaleMode.None; // children may scale, but we won't interfere.
-                mouseHook = new MouseHook(this, this, psheet);
-                Visible = false;
-                gridView = psheet;
-                this.BackColor = gridView.BackColor;
-            }
-
-            protected override CreateParams CreateParams {
-                get {
-                    CreateParams cp = base.CreateParams;
-                    cp.ExStyle |= NativeMethods.WS_EX_TOOLWINDOW;
-                    cp.Style |= NativeMethods.WS_POPUP | NativeMethods.WS_BORDER;
-                    if (OSFeature.IsPresent(SystemParameter.DropShadow))
-                    {
-                        cp.ClassStyle |= (int)NativeMethods.ClassStyle.CS_DROPSHADOW;
-                    }
-                    if (gridView != null)
-                    {
-                        cp.Parent = gridView.ParentInternal.Handle;
-                    }
-                    return cp;
-                }
-            }
-
-            private LinkLabel CreateNewLink
-            {
-                get
-                {
-                    if (createNewLink == null)
-                    {
-                        createNewLink = new LinkLabel();
-                        createNewLink.LinkClicked += new LinkLabelLinkClickedEventHandler(OnNewLinkClicked);
-                    }
-                    return createNewLink;
-                }
-            }
-
-            public virtual bool HookMouseDown
-            {
-                get
-                {
-                    return mouseHook.HookMouseDown;
-                }
-                set
-                {
-                    mouseHook.HookMouseDown = value;
-                }
-            }
-
-            /// <summary>
-            ///  This gets set to true if there isn't enough space below the currently selected
-            ///  row for the drop down, so it appears above the row.  In this case, we make the resize
-            ///  grip appear at the top left.
-            /// </summary>
-            public bool ResizeUp
-            {
-
-                set
-                {
-                    if (resizeUp != value)
-                    {
-
-                        // clear the glyph so we regenerate it.
-                        //
-                        sizeGripGlyph = null;
-                        resizeUp = value;
-
-                        if (resizable)
-                        {
-                            DockPadding.Bottom = 0;
-                            DockPadding.Top = 0;
-                            if (value)
-                            {
-                                DockPadding.Top = ResizeBarSize;
-                            }
-                            else
-                            {
-                                DockPadding.Bottom = ResizeBarSize;
-                            }
-                        }
-                    }
-                }
-            }
-
-            protected override void DestroyHandle()
-            {
-                mouseHook.HookMouseDown = false;
-                base.DestroyHandle();
-            }
-
-            protected override void Dispose(bool disposing)
-            {
-
-                if (disposing && createNewLink != null)
-                {
-                    createNewLink.Dispose();
-                    createNewLink = null;
-                }
-                base.Dispose(disposing);
-            }
-
-            public void DoModalLoop()
-            {
-
-                // Push a modal loop.  This seems expensive, but I think it is a
-                // better user model than returning from DropDownControl immediately.
-                //
-                while (Visible)
-                {
-                    Application.DoEventsModal();
-                    UnsafeNativeMethods.MsgWaitForMultipleObjectsEx(0, IntPtr.Zero, 250, NativeMethods.QS_ALLINPUT, NativeMethods.MWMO_INPUTAVAILABLE);
-                }
-            }
-
-            public virtual Control Component
-            {
-                get
-                {
-                    return currentControl;
-                }
-            }
-
-            /// <summary>
-            ///  Get an InstanceCreationEditor for this entry.  First, we look on the property type, and if we
-            ///  don't find that we'll go up to the editor type itself.  That way people can associate the InstanceCreationEditor with
-            ///  the type of DropDown UIType Editor.
-            ///
-            /// </summary>
-            private InstanceCreationEditor GetInstanceCreationEditor(PropertyDescriptorGridEntry entry)
-            {
-
-                if (entry == null)
-                {
-                    return null;
-                }
-
-                InstanceCreationEditor editor = null;
-
-                // check the property type itself.  this is the default path.
-                //
-                PropertyDescriptor pd = entry.PropertyDescriptor;
-                if (pd != null)
-                {
-                    editor = pd.GetEditor(typeof(InstanceCreationEditor)) as InstanceCreationEditor;
-                }
-
-                // now check if there is a dropdown UI type editor.  If so, use that.
-                //
-                if (editor == null)
-                {
-                    UITypeEditor ute = entry.UITypeEditor;
-                    if (ute != null && ute.GetEditStyle() == UITypeEditorEditStyle.DropDown)
-                    {
-                        editor = (InstanceCreationEditor)TypeDescriptor.GetEditor(ute, typeof(InstanceCreationEditor));
-                    }
-                }
-                return editor;
-            }
-
-            /// <summary>
-            ///  Get a glyph for sizing the lower left hand grip.  The code in ControlPaint only does lower-right glyphs
-            ///  so we do some GDI+ magic to take that glyph and mirror it.  That way we can still share the code (in case it changes for theming, etc),
-            ///  not have any special cases, and possibly solve world hunger.
-            /// </summary>
-            private Bitmap GetSizeGripGlyph(Graphics g)
-            {
-                if (sizeGripGlyph != null)
-                {
-                    return sizeGripGlyph;
-                }
-
-                // create our drawing surface based on the current graphics context.
-                //
-                sizeGripGlyph = new Bitmap(ResizeGripSize, ResizeGripSize, g);
-
-                using (Graphics glyphGraphics = Graphics.FromImage(sizeGripGlyph))
-                {
-                    // mirror the image around the x-axis to get a gripper handle that works
-                    // for the lower left.
-                    Matrix m = new Matrix();
-
-                    // basically, mirroring is just scaling by -1 on the X-axis.  So any point that's like (10, 10) goes to (-10, 10).
-                    // that mirrors it, but also moves everything to the negative axis, so we just bump the whole thing over by it's width.
-                    //
-                    // the +1 is because things at (0,0) stay at (0,0) since [0 * -1 = 0] so we want to get them over to the other side too.
-                    //
-                    // resizeUp causes the image to also be mirrored vertically so the grip can be used as a top-left grip instead of bottom-left.
-                    //
-                    m.Translate(ResizeGripSize + 1, (resizeUp ? ResizeGripSize + 1 : 0));
-                    m.Scale(-1, (resizeUp ? -1 : 1));
-                    glyphGraphics.Transform = m;
-                    ControlPaint.DrawSizeGrip(glyphGraphics, BackColor, 0, 0, ResizeGripSize, ResizeGripSize);
-                    glyphGraphics.ResetTransform();
-
-                }
-                sizeGripGlyph.MakeTransparent(BackColor);
-                return sizeGripGlyph;
-            }
-
-            public virtual bool GetUsed()
-            {
-                return (currentControl != null);
-            }
-
-            public virtual void FocusComponent()
-            {
-                Debug.WriteLineIf(CompModSwitches.DebugGridView.TraceVerbose, "DropDownHolder:FocusComponent()");
-                if (currentControl != null && Visible)
-                {
-                    currentControl.Focus();
-                }
-            }
-
-            /// <summary>
-            ///  General purpose method, based on Control.Contains()...
-            ///
-            ///  Determines whether a given window (specified using native window handle)
-            ///  is a descendant of this control. This catches both contained descendants
-            ///  and 'owned' windows such as modal dialogs. Using window handles rather
-            ///  than Control objects allows it to catch un-managed windows as well.
-            /// </summary>
-            private bool OwnsWindow(IntPtr hWnd)
-            {
-                while (hWnd != IntPtr.Zero)
-                {
-                    hWnd = UnsafeNativeMethods.GetWindowLong(new HandleRef(null, hWnd), NativeMethods.GWL_HWNDPARENT);
-                    if (hWnd == IntPtr.Zero)
-                    {
-                        return false;
-                    }
-                    if (hWnd == Handle)
-                    {
-                        return true;
-                    }
-                }
-                return false;
-            }
-
-            public bool OnClickHooked()
-            {
-                gridView.CloseDropDownInternal(false);
-                return false;
-            }
-
-            private void OnCurrentControlResize(object o, EventArgs e)
-            {
-                if (currentControl != null && !resizing)
-                {
-                    int oldWidth = Width;
-                    Size newSize = new Size(2 * DropDownHolderBorder + currentControl.Width, 2 * DropDownHolderBorder + currentControl.Height);
-                    if (resizable)
-                    {
-                        newSize.Height += ResizeBarSize;
-                    }
-                    try
-                    {
-                        resizing = true;
-                        SuspendLayout();
-                        Size = newSize;
-                    }
-                    finally
-                    {
-                        resizing = false;
-                        ResumeLayout(false);
-                    }
-                    Left -= (Width - oldWidth);
-                }
-            }
-
-            protected override void OnLayout(LayoutEventArgs levent)
-            {
-                try
-                {
-                    resizing = true;
-                    base.OnLayout(levent);
-                }
-                finally
-                {
-                    resizing = false;
-                }
-
-            }
-
-            private void OnNewLinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
-            {
-                InstanceCreationEditor ice = e.Link.LinkData as InstanceCreationEditor;
-
-                Debug.Assert(ice != null, "How do we have a link without the InstanceCreationEditor?");
-                if (ice != null)
-                {
-                    Type createType = gridView.SelectedGridEntry.PropertyType;
-                    if (createType != null)
-                    {
-
-                        gridView.CloseDropDown();
-
-                        object newValue = ice.CreateInstance(gridView.SelectedGridEntry, createType);
-
-                        if (newValue != null)
-                        {
-
-                            // make sure we got what we asked for.
-                            //
-                            if (!createType.IsInstanceOfType(newValue))
-                            {
-                                throw new InvalidCastException(string.Format(SR.PropertyGridViewEditorCreatedInvalidObject, createType));
-                            }
-
-                            gridView.CommitValue(newValue);
-                        }
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Just figure out what kind of sizing we would do at a given drag location.
-            /// </summary>
-            private int MoveTypeFromPoint(int x, int y)
-            {
-                Rectangle bGripRect = new Rectangle(0, Height - ResizeGripSize, ResizeGripSize, ResizeGripSize);
-                Rectangle tGripRect = new Rectangle(0, 0, ResizeGripSize, ResizeGripSize);
-
-                if (!resizeUp && bGripRect.Contains(x, y))
-                {
-                    return MoveTypeLeft | MoveTypeBottom;
-                }
-                else if (resizeUp && tGripRect.Contains(x, y))
-                {
-                    return MoveTypeLeft | MoveTypeTop;
-                }
-                else if (!resizeUp && Math.Abs(Height - y) < ResizeBorderSize)
-                {
-                    return MoveTypeBottom;
-                }
-                else if (resizeUp && Math.Abs(y) < ResizeBorderSize)
-                {
-                    return MoveTypeTop;
-                }
-                return MoveTypeNone;
-            }
-
-            /// <summary>
-            ///  Decide if we're going to be sizing at the given point, and if so, Capture and safe our current state.
-            /// </summary>
-            protected override void OnMouseDown(MouseEventArgs e)
-            {
-                if (e.Button == MouseButtons.Left)
-                {
-                    currentMoveType = MoveTypeFromPoint(e.X, e.Y);
-                    if (currentMoveType != MoveTypeNone)
-                    {
-                        dragStart = PointToScreen(new Point(e.X, e.Y));
-                        dragBaseRect = Bounds;
-                        Capture = true;
-                    }
-                    else
-                    {
-                        gridView.CloseDropDown();
-                    }
-                }
-                base.OnMouseDown(e);
-            }
-
-            /// <summary>
-            ///  Either set the cursor or do a move, depending on what our currentMoveType is/
-            /// </summary>
-            protected override void OnMouseMove(MouseEventArgs e)
-            {
-                // not moving so just set the cursor.
-                //
-                if (currentMoveType == MoveTypeNone)
-                {
-                    int cursorMoveType = MoveTypeFromPoint(e.X, e.Y);
-                    switch (cursorMoveType)
-                    {
-                        case (MoveTypeLeft | MoveTypeBottom):
-                            Cursor = Cursors.SizeNESW;
-                            break;
-                        case MoveTypeBottom:
-                        case MoveTypeTop:
-                            Cursor = Cursors.SizeNS;
-                            break;
-                        case MoveTypeTop | MoveTypeLeft:
-                            Cursor = Cursors.SizeNWSE;
-                            break;
-                        default:
-                            Cursor = null;
-                            break;
-                    }
-                }
-                else
-                {
-                    Point dragPoint = PointToScreen(new Point(e.X, e.Y));
-                    Rectangle newBounds = Bounds;
-
-                    // we're in a move operation, so do the resize.
-                    //
-                    if ((currentMoveType & MoveTypeBottom) == MoveTypeBottom)
-                    {
-                        newBounds.Height = Math.Max(MinDropDownSize.Height, dragBaseRect.Height + (dragPoint.Y - dragStart.Y));
-                    }
-
-                    // for left and top moves, we actually have to resize and move the form simultaneously.
-                    // do to that, we compute the xdelta, and apply that to the base rectangle if it's not going to
-                    // make the form smaller than the minimum.
-                    //
-                    if ((currentMoveType & MoveTypeTop) == MoveTypeTop)
-                    {
-                        int delta = dragPoint.Y - dragStart.Y;
-
-                        if ((dragBaseRect.Height - delta) > MinDropDownSize.Height)
-                        {
-                            newBounds.Y = dragBaseRect.Top + delta;
-                            newBounds.Height = dragBaseRect.Height - delta;
-                        }
-                    }
-
-                    if ((currentMoveType & MoveTypeLeft) == MoveTypeLeft)
-                    {
-                        int delta = dragPoint.X - dragStart.X;
-
-                        if ((dragBaseRect.Width - delta) > MinDropDownSize.Width)
-                        {
-                            newBounds.X = dragBaseRect.Left + delta;
-                            newBounds.Width = dragBaseRect.Width - delta;
-                        }
-                    }
-
-                    if (newBounds != Bounds)
-                    {
-                        try
-                        {
-                            resizing = true;
-                            Bounds = newBounds;
-                        }
-                        finally
-                        {
-                            resizing = false;
-                        }
-                    }
-
-                    // Redraw!
-                    //
-                    Invalidate();
-                }
-                base.OnMouseMove(e);
-            }
-
-            protected override void OnMouseLeave(EventArgs e)
-            {
-                // just clear the cursor back to the default.
-                //
-                Cursor = null;
-                base.OnMouseLeave(e);
-            }
-
-            protected override void OnMouseUp(MouseEventArgs e)
-            {
-                base.OnMouseUp(e);
-
-                if (e.Button == MouseButtons.Left)
-                {
-
-                    // reset the world.
-                    //
-                    currentMoveType = MoveTypeNone;
-                    dragStart = Point.Empty;
-                    dragBaseRect = Rectangle.Empty;
-                    Capture = false;
-                }
-            }
-
-            /// <summary>
-            ///  Just paint and draw our glyph.
-            /// </summary>
-            protected override void OnPaint(PaintEventArgs pe)
-            {
-                base.OnPaint(pe);
-                if (resizable)
-                {
-                    // Draw the grip
-                    Rectangle lRect = new Rectangle(0, resizeUp ? 0 : Height - ResizeGripSize, ResizeGripSize, ResizeGripSize);
-                    pe.Graphics.DrawImage(GetSizeGripGlyph(pe.Graphics), lRect);
-
-                    // Draw the divider
-                    int y = resizeUp ? (ResizeBarSize - 1) : (Height - ResizeBarSize);
-                    Pen pen = new Pen(SystemColors.ControlDark, 1)
-                    {
-                        DashStyle = DashStyle.Solid
-                    };
-                    pe.Graphics.DrawLine(pen, 0, y, Width, y);
-                    pen.Dispose();
-                }
-            }
-
-            protected override bool ProcessDialogKey(Keys keyData)
-            {
-                if ((keyData & (Keys.Shift | Keys.Control | Keys.Alt)) == 0)
-                {
-                    switch (keyData & Keys.KeyCode)
-                    {
-                        case Keys.Escape:
-                            gridView.OnEscape(this);
-                            return true;
-                        case Keys.F4:
-                            gridView.F4Selection(true);
-                            return true;
-                        case Keys.Return:
-                            // make sure the return gets forwarded to the control that
-                            // is being displayed
-                            if (gridView.UnfocusSelection() && gridView.SelectedGridEntry != null)
-                            {
-                                gridView.SelectedGridEntry.OnValueReturnKey();
-                            }
-                            return true;
-                    }
-                }
-
-                return base.ProcessDialogKey(keyData);
-            }
-
-            public void SetComponent(Control ctl, bool resizable)
-            {
-
-                this.resizable = resizable;
-                Font = gridView.Font;
-
-                // check to see if we're going to be adding an InstanceCreationEditor
-                //
-                InstanceCreationEditor editor = (ctl == null ? null : GetInstanceCreationEditor(gridView.SelectedGridEntry as PropertyDescriptorGridEntry));
-
-                // clear any existing control we have
-                //
-                if (currentControl != null)
-                {
-                    currentControl.Resize -= new EventHandler(OnCurrentControlResize);
-                    Controls.Remove(currentControl);
-                    currentControl = null;
-                }
-
-                // remove the InstanceCreationEditor link
-                //
-                if (createNewLink != null && createNewLink.Parent == this)
-                {
-                    Controls.Remove(createNewLink);
-                }
-
-                // now set up the new control, top to bottom
-                //
-                if (ctl != null)
-                {
-
-                    currentControl = ctl;
-                    Debug.WriteLineIf(CompModSwitches.DebugGridView.TraceVerbose, "DropDownHolder:SetComponent(" + (ctl.GetType().Name) + ")");
-
-                    DockPadding.All = 0;
-
-                    // first handle the control.  If it's a listbox, make sure it's got some height
-                    // to it.
-                    //
-                    if (currentControl is GridViewListBox)
-                    {
-                        ListBox lb = (ListBox)currentControl;
-
-                        if (lb.Items.Count == 0)
-                        {
-                            lb.Height = Math.Max(lb.Height, lb.ItemHeight);
-                        }
-                    }
-
-                    // Parent the control now.  That way it can inherit our
-                    // font and scale itself if it wants to.
-                    try
-                    {
-                        SuspendLayout();
-                        Controls.Add(ctl);
-
-                        Size sz = new Size(2 * DropDownHolderBorder + ctl.Width, 2 * DropDownHolderBorder + ctl.Height);
-
-                        // now check for an editor, and show the link if there is one.
-                        //
-                        if (editor != null)
-                        {
-
-                            // set up the link.
-                            //
-                            CreateNewLink.Text = editor.Text;
-                            CreateNewLink.Links.Clear();
-                            CreateNewLink.Links.Add(0, editor.Text.Length, editor);
-
-                            // size it as close to the size of the text as possible.
-                            //
-                            int linkHeight = CreateNewLink.Height;
-                            using (Graphics g = gridView.CreateGraphics())
-                            {
-                                SizeF sizef = PropertyGrid.MeasureTextHelper.MeasureText(gridView.ownerGrid, g, editor.Text, gridView.GetBaseFont());
-                                linkHeight = (int)sizef.Height;
-                            }
-
-                            CreateNewLink.Height = linkHeight + DropDownHolderBorder;
-
-                            // add the total height plus some border
-                            sz.Height += (linkHeight + (DropDownHolderBorder * 2));
-                        }
-
-                        // finally, if we're resizable, add the space for the widget.
-                        //
-                        if (resizable)
-                        {
-                            sz.Height += ResizeBarSize;
-
-                            // we use dockpadding to save space to draw the widget.
-                            //
-                            if (resizeUp)
-                            {
-                                DockPadding.Top = ResizeBarSize;
-                            }
-                            else
-                            {
-                                DockPadding.Bottom = ResizeBarSize;
-                            }
-                        }
-
-                        // set the size stuff.
-                        //
-                        Size = sz;
-
-                        // DockStyle property on child control is forcing to resize to Parent control size that is not yet scaled.
-                        // Parent control's Preffered size, by this time, is calculated based on child controls scaled dimensions.
-                        // Also, making sure we are not shrinking if we are already there.
-
-                        ctl.Visible = true;
-                        if (this.Size.Height < this.PreferredSize.Height)  {
-                            this.Size = new Size(this.Size.Width, this.PreferredSize.Height);
-                        }
-                        ctl.Dock = DockStyle.Fill;
-
-                        if (editor != null)
-                        {
-                            CreateNewLink.Dock = DockStyle.Bottom;
-                            Controls.Add(CreateNewLink);
-                        }
-                    }
-                    finally
-                    {
-                        ResumeLayout(true);
-                    }
-
-                    // hook the resize event.
-                    //
-                    currentControl.Resize += new EventHandler(OnCurrentControlResize);
-                }
-                Enabled = currentControl != null;
-            }
-
-            protected override void WndProc(ref Message m)
-            {
-
-                if (m.Msg == WindowMessages.WM_ACTIVATE)
-                {
-                    SetState(STATE_MODAL, true);
-                    Debug.WriteLineIf(CompModSwitches.DebugGridView.TraceVerbose, "DropDownHolder:WM_ACTIVATE()");
-                    IntPtr activatedWindow = (IntPtr)m.LParam;
-                    if (Visible && NativeMethods.Util.LOWORD(m.WParam) == NativeMethods.WA_INACTIVE && !OwnsWindow(activatedWindow))
-                    {
-                        gridView.CloseDropDownInternal(false);
-                        return;
-                    }
-
-                    // prevent the IMsoComponentManager active code from getting fired.
-                    //Active = ((int)m.WParam & 0x0000FFFF) != NativeMethods.WA_INACTIVE;
-                    //return;
-                }
-                else if (m.Msg == WindowMessages.WM_CLOSE)
-                {
-                    // don't let an ALT-F4 get you down
-                    //
-                    if (Visible)
-                    {
-                        gridView.CloseDropDown();
-                    }
-                    return;
-                }
-                else if (m.Msg == WindowMessages.WM_DPICHANGED)
-                {
-                    // Dropdownholder in PropertyGridView is already scaled based on parent font and other properties that were already set for new DPI
-                    // This case is to avoid rescaling(double scaling) of this form
-
-                    var oldDpi = this.deviceDpi;
-                    this.deviceDpi =  (int)UnsafeNativeMethods.GetDpiForWindow(new HandleRef(this, HandleInternal));
-                    if (oldDpi != deviceDpi)  {
-                        RescaleConstantsForDpi(oldDpi, this.deviceDpi);
-                        this.PerformLayout();
-                    }
-
-                    m.Result = IntPtr.Zero;
-                    return;
-                }
-
-                base.WndProc(ref m);
-            }
-
-            protected override void RescaleConstantsForDpi(int oldDpi, int newDpi) {
-                base.RescaleConstantsForDpi(oldDpi, newDpi);
-
-                var scrollbarHeight = SystemInformation.GetHorizontalScrollBarHeightForDpi(newDpi);
-                MinDropDownSize = new Size(SystemInformation.GetVerticalScrollBarWidthForDpi(newDpi) * 4, scrollbarHeight * 4);
-                ResizeGripSize = scrollbarHeight;
-
-                ResizeBarSize = ResizeGripSize + 1;
-                ResizeBorderSize = ResizeBarSize / 2;
-
-                var factor = (double)newDpi / oldDpi;
-                Height = (int)Math.Round(factor * Height);
-            }
-        }
-
-        private class GridViewListBox : ListBox
-        {
-            internal bool fInSetSelectedIndex = false;
-            private readonly PropertyGridView _owningPropertyGridView;
-
-            public GridViewListBox(PropertyGridView gridView)
-            {
-                base.IntegralHeight = false;
-                _owningPropertyGridView = gridView;
-                base.BackColor = gridView.BackColor;
-            }
-
-            protected override CreateParams CreateParams
-            {
-                get
-                {
-                    CreateParams cp = base.CreateParams;
-                    cp.Style &= ~NativeMethods.WS_BORDER;
-                    cp.ExStyle &= ~NativeMethods.WS_EX_CLIENTEDGE;
-                    return cp;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the owning PropertyGridView.
-            /// </summary>
-            internal PropertyGridView OwningPropertyGridView
-            {
-                get
-                {
-                    return _owningPropertyGridView;
-                }
-            }
-
-            /// <summary>
-            ///  Indicates whether or not the control supports UIA Providers via
-            ///  IRawElementProviderFragment/IRawElementProviderFragmentRoot interfaces
-            /// </summary>
-            internal override bool SupportsUiaProviders => true;
-
-            /// <summary>
-            ///  Constructs the new instance of the accessibility object for this control.
-            /// </summary>
-            /// <returns>The accessibility object instance.</returns>
-            protected override AccessibleObject CreateAccessibilityInstance()
-            {
-                return new GridViewListBoxAccessibleObject(this);
-            }
-
-            public virtual bool InSetSelectedIndex()
-            {
-                return fInSetSelectedIndex;
-            }
-
-            protected override void OnSelectedIndexChanged(EventArgs e)
-            {
-                fInSetSelectedIndex = true;
-                base.OnSelectedIndexChanged(e);
-                fInSetSelectedIndex = false;
-
-                if (AccessibilityObject is GridViewListBoxAccessibleObject gridViewListBoxAccessibleObject)
-                {
-                    gridViewListBoxAccessibleObject.SetListBoxItemFocus();
-                }
-            }
-
-        }
-
-        [ComVisible(true)]
-        private class GridViewListBoxItemAccessibleObject : AccessibleObject
-        {
-            private readonly GridViewListBox _owningGridViewListBox;
-            private readonly object _owningItem;
-
-            public GridViewListBoxItemAccessibleObject(GridViewListBox owningGridViewListBox, object owningItem)
-            {
-                _owningGridViewListBox = owningGridViewListBox;
-                _owningItem = owningItem;
-
-                UseStdAccessibleObjects(_owningGridViewListBox.Handle);
-            }
-
-            /// <summary>
-            ///  Gets the DropDown button bounds.
-            /// </summary>
-            public override Rectangle Bounds
-            {
-                get
-                {
-                    var systemIAccessible = GetSystemIAccessibleInternal();
-                    systemIAccessible.accLocation(out int left, out int top, out int width, out int height, GetChildId());
-                    return new Rectangle(left, top, width, height);
-                }
-            }
-
-            /// <summary>
-            ///  Gets the DropDown button default action.
-            /// </summary>
-            public override string DefaultAction
-            {
-                get
-                {
-                    var systemIAccessible = GetSystemIAccessibleInternal();
-                    return systemIAccessible.accDefaultAction[GetChildId()];
-                }
-            }
-
-            internal override UnsafeNativeMethods.IRawElementProviderFragment FragmentNavigate(UnsafeNativeMethods.NavigateDirection direction)
-            {
-                switch (direction)
-                {
-                    case UnsafeNativeMethods.NavigateDirection.Parent:
-                        return _owningGridViewListBox.AccessibilityObject;
-                    case UnsafeNativeMethods.NavigateDirection.NextSibling:
-                        int currentIndex = GetCurrentIndex();
-                        if (_owningGridViewListBox.AccessibilityObject is GridViewListBoxAccessibleObject gridViewListBoxAccessibleObject)
-                        {
-                            int itemsCount = gridViewListBoxAccessibleObject.GetChildFragmentCount();
-                            int nextItemIndex = currentIndex + 1;
-                            if (itemsCount > nextItemIndex)
-                            {
-                                return gridViewListBoxAccessibleObject.GetChildFragment(nextItemIndex);
-                            }
-                        }
-                        break;
-                    case UnsafeNativeMethods.NavigateDirection.PreviousSibling:
-                        currentIndex = GetCurrentIndex();
-                        gridViewListBoxAccessibleObject = _owningGridViewListBox.AccessibilityObject as GridViewListBoxAccessibleObject;
-                        if (gridViewListBoxAccessibleObject != null)
-                        {
-                            var itemsCount = gridViewListBoxAccessibleObject.GetChildFragmentCount();
-                            int previousItemIndex = currentIndex - 1;
-                            if (previousItemIndex >= 0)
-                            {
-                                return gridViewListBoxAccessibleObject.GetChildFragment(previousItemIndex);
-                            }
-                        }
-
-                        break;
-                }
-
-                return base.FragmentNavigate(direction);
-            }
-
-            internal override UnsafeNativeMethods.IRawElementProviderFragmentRoot FragmentRoot
-            {
-                get
-                {
-                    return _owningGridViewListBox.AccessibilityObject;
-                }
-            }
-
-            private int GetCurrentIndex()
-            {
-                return _owningGridViewListBox.Items.IndexOf(_owningItem);
-            }
-
-            internal override int GetChildId()
-            {
-                return GetCurrentIndex() + 1; // Index is zero-based, Child ID is 1-based.
-            }
-
-            internal override object GetPropertyValue(int propertyID)
-            {
-                switch (propertyID)
-                {
-                    case NativeMethods.UIA_RuntimeIdPropertyId:
-                        return RuntimeId;
-                    case NativeMethods.UIA_BoundingRectanglePropertyId:
-                        return BoundingRectangle;
-                    case NativeMethods.UIA_ControlTypePropertyId:
-                        return NativeMethods.UIA_ListItemControlTypeId;
-                    case NativeMethods.UIA_NamePropertyId:
-                        return Name;
-                    case NativeMethods.UIA_AccessKeyPropertyId:
-                        return KeyboardShortcut;
-                    case NativeMethods.UIA_HasKeyboardFocusPropertyId:
-                        return _owningGridViewListBox.Focused;
-                    case NativeMethods.UIA_IsKeyboardFocusablePropertyId:
-                        return (State & AccessibleStates.Focusable) == AccessibleStates.Focusable;
-                    case NativeMethods.UIA_IsEnabledPropertyId:
-                        return _owningGridViewListBox.Enabled;
-                    case NativeMethods.UIA_HelpTextPropertyId:
-                        return Help ?? string.Empty;
-                    case NativeMethods.UIA_IsPasswordPropertyId:
-                        return false;
-                    case NativeMethods.UIA_IsOffscreenPropertyId:
-                        return (State & AccessibleStates.Offscreen) == AccessibleStates.Offscreen;
-                    default:
-                        return base.GetPropertyValue(propertyID);
-                }
-            }
-
-            /// <summary>
-            ///  Gets the help text.
-            /// </summary>
-            public override string Help
-            {
-                get
-                {
-                    var systemIAccessible = GetSystemIAccessibleInternal();
-                    return systemIAccessible.accHelp[GetChildId()];
-                }
-            }
-
-            /// <summary>
-            ///  Gets the keyboard shortcut.
-            /// </summary>
-            public override string KeyboardShortcut
-            {
-                get
-                {
-                    var systemIAccessible = GetSystemIAccessibleInternal();
-                    return systemIAccessible.get_accKeyboardShortcut(GetChildId());
-                }
-            }
-
-            /// <summary>
-            ///  Indicates whether specified pattern is supported.
-            /// </summary>
-            /// <param name="patternId">The pattern ID.</param>
-            /// <returns>True if specified </returns>
-            internal override bool IsPatternSupported(int patternId)
-            {
-                if (patternId == NativeMethods.UIA_LegacyIAccessiblePatternId ||
-                    patternId == NativeMethods.UIA_InvokePatternId)
-                {
-                    return true;
-                }
-
-                return base.IsPatternSupported(patternId);
-            }
-
-            /// <summary>
-            ///  Gets or sets the accessible name.
-            /// </summary>
-            public override string Name
-            {
-                get
-                {
-                    if (_owningGridViewListBox != null)
-                    {
-                        return _owningItem.ToString();
-                    }
-
-                    return base.Name;
-                }
-
-                set
-                {
-                    base.Name = value;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the accessible role.
-            /// </summary>
-            public override AccessibleRole Role
-            {
-                get
-                {
-                    var systemIAccessible = GetSystemIAccessibleInternal();
-                    return (AccessibleRole)systemIAccessible.get_accRole(GetChildId());
-                }
-            }
-
-            /// <summary>
-            ///  Gets the runtime ID.
-            /// </summary>
-            internal override int[] RuntimeId
-            {
-                get
-                {
-                    var runtimeId = new int[3];
-                    runtimeId[0] = RuntimeIDFirstItem;
-                    runtimeId[1] = (int)(long)_owningGridViewListBox.Handle;
-                    runtimeId[2] = _owningItem.GetHashCode();
-
-                    return runtimeId;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the accessible state.
-            /// </summary>
-            public override AccessibleStates State
-            {
-                get
-                {
-                    var systemIAccessible = GetSystemIAccessibleInternal();
-                    return (AccessibleStates)systemIAccessible.get_accState(GetChildId());
-                }
-            }
-
-            internal override void SetFocus()
-            {
-                RaiseAutomationEvent(NativeMethods.UIA_AutomationFocusChangedEventId);
-
-                base.SetFocus();
-            }
-        }
-
-        private class GridViewListBoxItemAccessibleObjectCollection : Hashtable
-        {
-            private readonly GridViewListBox _owningGridViewListBox;
-
-            public GridViewListBoxItemAccessibleObjectCollection(GridViewListBox owningGridViewListBox)
-            {
-                _owningGridViewListBox = owningGridViewListBox;
-            }
-
-            public override object this[object key]
-            {
-                get
-                {
-                    if (!ContainsKey(key))
-                    {
-                        var itemAccessibleObject = new GridViewListBoxItemAccessibleObject(_owningGridViewListBox, key);
-                        base[key] = itemAccessibleObject;
-                    }
-
-                    return base[key];
-                }
-
-                set
-                {
-                    base[key] = value;
-                }
-            }
-        }
-
-        /// <summary>
-        ///  Represents the PropertyGridView ListBox accessibility object.
-        /// </summary>
-        [ComVisible(true)]
-        private class GridViewListBoxAccessibleObject : ControlAccessibleObject
-        {
-            private readonly GridViewListBox _owningGridViewListBox;
-            private readonly PropertyGridView _owningPropertyGridView;
-            private readonly GridViewListBoxItemAccessibleObjectCollection _itemAccessibleObjects;
-
-            /// <summary>
-            ///  Constructs the new instance of GridViewListBoxAccessibleObject.
-            /// </summary>
-            /// <param name="owningGridViewListBox">The owning GridViewListBox.</param>
-            public GridViewListBoxAccessibleObject(GridViewListBox owningGridViewListBox) : base(owningGridViewListBox)
-            {
-                _owningGridViewListBox = owningGridViewListBox;
-                _owningPropertyGridView = owningGridViewListBox.OwningPropertyGridView;
-                _itemAccessibleObjects = new GridViewListBoxItemAccessibleObjectCollection(owningGridViewListBox);
-            }
-
-            /// <summary>
-            ///  Request to return the element in the specified direction.
-            /// </summary>
-            /// <param name="direction">Indicates the direction in which to navigate.</param>
-            /// <returns>Returns the element in the specified direction.</returns>
-            internal override UnsafeNativeMethods.IRawElementProviderFragment FragmentNavigate(UnsafeNativeMethods.NavigateDirection direction)
-            {
-                if (direction == UnsafeNativeMethods.NavigateDirection.Parent)
-                {
-                    return _owningPropertyGridView.SelectedGridEntry.AccessibilityObject;
-                }
-                else if (direction == UnsafeNativeMethods.NavigateDirection.FirstChild)
-                {
-                    return GetChildFragment(0);
-                }
-                else if (direction == UnsafeNativeMethods.NavigateDirection.LastChild)
-                {
-                    var childFragmentCount = GetChildFragmentCount();
-                    if (childFragmentCount > 0)
-                    {
-                        return GetChildFragment(childFragmentCount - 1);
-                    }
-                }
-                else if (direction == UnsafeNativeMethods.NavigateDirection.NextSibling)
-                {
-                    return _owningPropertyGridView.Edit.AccessibilityObject;
-                }
-
-                return base.FragmentNavigate(direction);
-            }
-
-            /// <summary>
-            ///  Return the element that is the root node of this fragment of UI.
-            /// </summary>
-            internal override UnsafeNativeMethods.IRawElementProviderFragmentRoot FragmentRoot
-            {
-                get
-                {
-                    return _owningPropertyGridView.AccessibilityObject;
-                }
-            }
-
-            public AccessibleObject GetChildFragment(int index)
-            {
-                if (index < 0 || index >= _owningGridViewListBox.Items.Count)
-                {
-                    return null;
-                }
-
-                var item = _owningGridViewListBox.Items[index];
-                return _itemAccessibleObjects[item] as AccessibleObject;
-            }
-
-            public int GetChildFragmentCount()
-            {
-                return _owningGridViewListBox.Items.Count;
-            }
-
-            /// <summary>
-            ///  Request value of specified property from an element.
-            /// </summary>
-            /// <param name="propertyId">Identifier indicating the property to return</param>
-            /// <returns>Returns a ValInfo indicating whether the element supports this property, or has no value for it.</returns>
-            internal override object GetPropertyValue(int propertyID)
-            {
-                if (propertyID == NativeMethods.UIA_ControlTypePropertyId)
-                {
-                    return NativeMethods.UIA_ListControlTypeId;
-                }
-                else if (propertyID == NativeMethods.UIA_NamePropertyId)
-                {
-                    return Name;
-                }
-
-                return base.GetPropertyValue(propertyID);
-            }
-
-            internal override void SetFocus()
-            {
-                RaiseAutomationEvent(NativeMethods.UIA_AutomationFocusChangedEventId);
-
-                base.SetFocus();
-            }
-
-            internal void SetListBoxItemFocus()
-            {
-                var selectedItem = _owningGridViewListBox.SelectedItem;
-                if (_itemAccessibleObjects[selectedItem] is AccessibleObject itemAccessibleObject)
-                {
-                    itemAccessibleObject.SetFocus();
-                }
-            }
-        }
-
-        private class GridViewEdit : TextBox, IMouseHookClient
-        {
-            internal bool fInSetText = false;
-            internal bool filter = false;
-            internal PropertyGridView psheet = null;
-            private bool dontFocusMe = false;
-            private int lastMove;
-
-            private readonly MouseHook mouseHook;
-
-            // We do this becuase the Focus call above doesn't always stick, so
-            // we make the Edit think that it doesn't have focus.  this prevents
-            // ActiveControl code on the containercontrol from moving focus elsewhere
-            // when the dropdown closes.
-            public bool DontFocus
-            {
-                set
-                {
-                    dontFocusMe = value;
-                }
-            }
-
-            public virtual bool Filter
-            {
-                get { return filter; }
-
-                set
-                {
-                    filter = value;
-                }
-            }
-
-            /// <summary>
-            ///  Indicates whether or not the control supports UIA Providers via
-            ///  IRawElementProviderFragment/IRawElementProviderFragmentRoot interfaces
-            /// </summary>
-            internal override bool SupportsUiaProviders => true;
-
-            public override bool Focused
-            {
-                get
-                {
-                    if (dontFocusMe)
-                    {
-                        return false;
-                    }
-                    return base.Focused;
-                }
-            }
-
-            public override string Text
-            {
-                get
-                {
-                    return base.Text;
-                }
-                set
-                {
-                    fInSetText = true;
-                    base.Text = value;
-                    fInSetText = false;
-                }
-            }
-
-            public bool DisableMouseHook
-            {
-
-                set
-                {
-                    mouseHook.DisableMouseHook = value;
-                }
-            }
-
-            public virtual bool HookMouseDown
-            {
-                get
-                {
-                    return mouseHook.HookMouseDown;
-                }
-                set
-                {
-                    mouseHook.HookMouseDown = value;
-                    if (value)
-                    {
-                        Focus();
-                    }
-                }
-            }
-
-            public GridViewEdit(PropertyGridView psheet)
-            {
-                this.psheet = psheet;
-                mouseHook = new MouseHook(this, this, psheet);
-            }
-
-            /// <summary>
-            ///  Creates a new AccessibleObject for this GridViewEdit instance.
-            ///  The AccessibleObject instance returned by this method overrides several UIA properties.
-            /// </summary>
-            /// <returns>
-            ///  AccessibleObject for this GridViewEdit instance.
-            /// </returns>
-            protected override AccessibleObject CreateAccessibilityInstance()
-            {
-                return new GridViewEditAccessibleObject(this);
-            }
-
-            protected override void DestroyHandle()
-            {
-                mouseHook.HookMouseDown = false;
-                base.DestroyHandle();
-            }
-
-            protected override void Dispose(bool disposing)
-            {
-                if (disposing)
-                {
-                    mouseHook.Dispose();
-                }
-                base.Dispose(disposing);
-            }
-
-            public void FilterKeyPress(char keyChar)
-            {
-
-                if (IsInputChar(keyChar))
-                {
-                    Focus();
-                    SelectAll();
-                    UnsafeNativeMethods.PostMessage(new HandleRef(this, Handle), WindowMessages.WM_CHAR, (IntPtr)keyChar, IntPtr.Zero);
-                }
-            }
-
-            /// <summary>
-            ///  Overridden to handle TAB key.
-            /// </summary>
-            protected override bool IsInputKey(Keys keyData)
-            {
-                switch (keyData & Keys.KeyCode)
-                {
-                    case Keys.Escape:
-                    case Keys.Tab:
-                    case Keys.F4:
-                    case Keys.F1:
-                    case Keys.Return:
-                        return false;
-                }
-                if (psheet.NeedsCommit)
-                {
-                    return false;
-                }
-                return base.IsInputKey(keyData);
-            }
-
-            /// <summary>
-            ///  Overridden to handle TAB key.
-            /// </summary>
-            protected override bool IsInputChar(char keyChar)
-            {
-                switch ((Keys)(int)keyChar)
-                {
-                    case Keys.Tab:
-                    case Keys.Return:
-                        return false;
-                }
-                return base.IsInputChar(keyChar);
-            }
-
-            protected override void OnKeyDown(KeyEventArgs ke)
-            {
-
-                // this is because on a dialog we may
-                // not get a chance to pre-process
-                //
-                if (ProcessDialogKey(ke.KeyData))
-                {
-                    ke.Handled = true;
-                    return;
-                }
-
-                base.OnKeyDown(ke);
-            }
-
-            protected override void OnKeyPress(KeyPressEventArgs ke)
-            {
-                if (!IsInputChar(ke.KeyChar))
-                {
-                    ke.Handled = true;
-                    return;
-                }
-                base.OnKeyPress(ke);
-            }
-
-            public bool OnClickHooked()
-            {
-                // can we commit this value?
-                // eat the value if we failed to commit.
-                return !psheet._Commit();
-            }
-
-            protected override void OnMouseEnter(EventArgs e)
-            {
-                base.OnMouseEnter(e);
-
-                if (!Focused)
-                {
-                    Graphics g = CreateGraphics();
-                    if (psheet.SelectedGridEntry != null &&
-                        ClientRectangle.Width <= psheet.SelectedGridEntry.GetValueTextWidth(Text, g, Font))
-                    {
-                        psheet.ToolTip.ToolTip = PasswordProtect ? "" : Text;
-                    }
-                    g.Dispose();
-                }
-
-            }
-
-            protected override bool ProcessCmdKey(ref Message msg, Keys keyData)
-            {
-
-                // make sure we allow the Edit to handle ctrl-z
-                switch (keyData & Keys.KeyCode)
-                {
-                    case Keys.Z:
-                    case Keys.C:
-                    case Keys.X:
-                    case Keys.V:
-                        if (
-                           ((keyData & Keys.Control) != 0) &&
-                           ((keyData & Keys.Shift) == 0) &&
-                           ((keyData & Keys.Alt) == 0))
-                        {
-                            return false;
-                        }
-                        break;
-
-                    case Keys.A:
-                        if (
-                           ((keyData & Keys.Control) != 0) &&
-                           ((keyData & Keys.Shift) == 0) &&
-                           ((keyData & Keys.Alt) == 0))
-                        {
-                            SelectAll();
-                            return true;
-                        }
-
-                        break;
-
-                    case Keys.Insert:
-                        if (((keyData & Keys.Alt) == 0))
-                        {
-                            if (((keyData & Keys.Control) != 0) ^ ((keyData & Keys.Shift) == 0))
-                            {
-                                return false;
-                            }
-                        }
-                        break;
-
-                    case Keys.Delete:
-                        if (
-                          ((keyData & Keys.Control) == 0) &&
-                          ((keyData & Keys.Shift) != 0) &&
-                          ((keyData & Keys.Alt) == 0))
-                        {
-                            return false;
-                        }
-                        else if (
-                              ((keyData & Keys.Control) == 0) &&
-                              ((keyData & Keys.Shift) == 0) &&
-                              ((keyData & Keys.Alt) == 0)
-                                )
-                        {
-                            // if this is just the delete key and we're on a non-text editable property that is resettable,
-                            // reset it now.
-                            //
-                            if (psheet.SelectedGridEntry != null && !psheet.SelectedGridEntry.Enumerable && !psheet.SelectedGridEntry.IsTextEditable && psheet.SelectedGridEntry.CanResetPropertyValue())
-                            {
-                                object oldValue = psheet.SelectedGridEntry.PropertyValue;
-                                psheet.SelectedGridEntry.ResetPropertyValue();
-                                psheet.UnfocusSelection();
-                                psheet.ownerGrid.OnPropertyValueSet(psheet.SelectedGridEntry, oldValue);
-                            }
-                        }
-                        break;
-                }
-                return base.ProcessCmdKey(ref msg, keyData);
-            }
-
-            /// <summary>
-            ///  Overrides Control.ProcessDialogKey to handle the Escape and Return
-            ///  keys.
-            /// </summary>
-            protected override bool ProcessDialogKey(Keys keyData)
-            {
-
-                // We don't do anything with modified keys here.
-                //
-                if ((keyData & (Keys.Shift | Keys.Control | Keys.Alt)) == 0)
-                {
-                    switch (keyData & Keys.KeyCode)
-                    {
-                        case Keys.Return:
-                            bool fwdReturn = !psheet.NeedsCommit;
-                            if (psheet.UnfocusSelection() && fwdReturn)
-                            {
-                                psheet.SelectedGridEntry.OnValueReturnKey();
-                            }
-                            return true;
-                        case Keys.Escape:
-                            psheet.OnEscape(this);
-                            return true;
-                        case Keys.F4:
-                            psheet.F4Selection(true);
-                            return true;
-                    }
-                }
-
-                // for the tab key, we want to commit before we allow it to be processed.
-                if ((keyData & Keys.KeyCode) == Keys.Tab && ((keyData & (Keys.Control | Keys.Alt)) == 0))
-                {
-                    return !psheet._Commit();
-                }
-
-                return base.ProcessDialogKey(keyData);
-            }
-
-            protected override void SetVisibleCore(bool value)
-            {
-                Debug.WriteLineIf(CompModSwitches.DebugGridView.TraceVerbose, "DropDownHolder:Visible(" + (value.ToString()) + ")");
-                // make sure we dont' have the mouse captured if
-                // we're going invisible
-                if (value == false && HookMouseDown)
-                {
-                    mouseHook.HookMouseDown = false;
-                }
-                base.SetVisibleCore(value);
-            }
-
-            // a mini version of process dialog key
-            // for responding to WM_GETDLGCODE
-            internal bool WantsTab(bool forward)
-            {
-                return psheet.WantsTab(forward);
-            }
-
-            private unsafe bool WmNotify(ref Message m)
-            {
-
-                if (m.LParam != IntPtr.Zero)
-                {
-                    NativeMethods.NMHDR* nmhdr = (NativeMethods.NMHDR*)m.LParam;
-
-                    if (nmhdr->hwndFrom == psheet.ToolTip.Handle)
-                    {
-                        switch (nmhdr->code)
-                        {
-                            case NativeMethods.TTN_SHOW:
-                                PropertyGridView.PositionTooltip(this, psheet.ToolTip, ClientRectangle);
-                                m.Result = (IntPtr)1;
-                                return true;
-                            default:
-                                psheet.WndProc(ref m);
-                                break;
-                        }
-                    }
-                }
-                return false;
-            }
-
-            protected override void WndProc(ref Message m)
-            {
-
-                if (filter)
-                {
-                    if (psheet.FilterEditWndProc(ref m))
-                    {
-                        return;
-                    }
-                }
-
-                switch (m.Msg)
-                {
-                    case WindowMessages.WM_STYLECHANGED:
-                        if ((unchecked((int)(long)m.WParam) & NativeMethods.GWL_EXSTYLE) != 0)
-                        {
-                            psheet.Invalidate();
-                        }
-                        break;
-                    case WindowMessages.WM_MOUSEMOVE:
-                        if (unchecked((int)(long)m.LParam) == lastMove)
-                        {
-                            return;
-                        }
-                        lastMove = unchecked((int)(long)m.LParam);
-                        break;
-                    case WindowMessages.WM_DESTROY:
-                        mouseHook.HookMouseDown = false;
-                        break;
-                    case WindowMessages.WM_SHOWWINDOW:
-                        if (IntPtr.Zero == m.WParam)
-                        {
-                            mouseHook.HookMouseDown = false;
-                        }
-                        break;
-                    case WindowMessages.WM_PASTE:
-                        /*if (!this.ReadOnly) {
-                            IDataObject dataObject = Clipboard.GetDataObject();
-                            Debug.Assert(dataObject != null, "Failed to get dataObject from clipboard");
-                            if (dataObject != null) {
-                                object data = dataObject.GetData(typeof(string));
-                                if (data != null) {
-                                    string clipboardText = data.ToString();
-                                    SelectedText = clipboardText;
-                                    m.result = 1;
-                                    return;
-                                }
-                            }
-                        }*/
-                        if (ReadOnly)
-                        {
-                            return;
-                        }
-                        break;
-
-                    case WindowMessages.WM_GETDLGCODE:
-
-                        m.Result = (IntPtr)((long)m.Result | NativeMethods.DLGC_WANTARROWS | NativeMethods.DLGC_WANTCHARS);
-                        if (psheet.NeedsCommit || WantsTab((ModifierKeys & Keys.Shift) == 0))
-                        {
-                            m.Result = (IntPtr)((long)m.Result | NativeMethods.DLGC_WANTALLKEYS | NativeMethods.DLGC_WANTTAB);
-                        }
-                        return;
-
-                    case WindowMessages.WM_NOTIFY:
-                        if (WmNotify(ref m))
-                        {
-                            return;
-                        }
-
-                        break;
-                }
-                base.WndProc(ref m);
-            }
-
-            public virtual bool InSetText()
-            {
-                return fInSetText;
-            }
-
-            [ComVisible(true)]
-            protected class GridViewEditAccessibleObject : ControlAccessibleObject
-            {
-
-                private readonly PropertyGridView propertyGridView;
-
-                public GridViewEditAccessibleObject(GridViewEdit owner) : base(owner)
-                {
-                    propertyGridView = owner.psheet;
-                }
-
-                public override AccessibleStates State
-                {
-                    get
-                    {
-                        AccessibleStates states = base.State;
-                        if (IsReadOnly)
-                        {
-                            states |= AccessibleStates.ReadOnly;
-                        }
-                        else
-                        {
-                            states &= ~AccessibleStates.ReadOnly;
-                        }
-                        return states;
-                    }
-                }
-
-                internal override bool IsIAccessibleExSupported() => true;
-
-                /// <summary>
-                ///  Returns the element in the specified direction.
-                /// </summary>
-                /// <param name="direction">Indicates the direction in which to navigate.</param>
-                /// <returns>Returns the element in the specified direction.</returns>
-                internal override UnsafeNativeMethods.IRawElementProviderFragment FragmentNavigate(UnsafeNativeMethods.NavigateDirection direction)
-                {
-                    if (direction == UnsafeNativeMethods.NavigateDirection.Parent)
-                    {
-                        return propertyGridView.SelectedGridEntry.AccessibilityObject;
-                    }
-                    else if (direction == UnsafeNativeMethods.NavigateDirection.NextSibling)
-                    {
-                        if (propertyGridView.DropDownButton.Visible)
-                        {
-                            return propertyGridView.DropDownButton.AccessibilityObject;
-                        }
-                        else if (propertyGridView.DialogButton.Visible)
-                        {
-                            return propertyGridView.DialogButton.AccessibilityObject;
-                        }
-                    }
-
-                    return base.FragmentNavigate(direction);
-                }
-
-                /// <summary>
-                ///  Gets the top level element.
-                /// </summary>
-                internal override UnsafeNativeMethods.IRawElementProviderFragmentRoot FragmentRoot
-                {
-                    get
-                    {
-                        return propertyGridView.AccessibilityObject;
-                    }
-                }
-
-                internal override object GetPropertyValue(int propertyID)
-                {
-                    if (propertyID == NativeMethods.UIA_IsEnabledPropertyId)
-                    {
-                        return !IsReadOnly;
-                    }
-                    else if (propertyID == NativeMethods.UIA_IsValuePatternAvailablePropertyId)
-                    {
-                        return IsPatternSupported(NativeMethods.UIA_ValuePatternId);
-                    }
-                    else if (propertyID == NativeMethods.UIA_ControlTypePropertyId)
-                    {
-                        return NativeMethods.UIA_EditControlTypeId;
-                    }
-                    else if (propertyID == NativeMethods.UIA_NamePropertyId)
-                    {
-                        return Name;
-                    }
-
-                    return base.GetPropertyValue(propertyID);
-                }
-
-                internal override bool IsPatternSupported(int patternId)
-                {
-                    if (patternId == NativeMethods.UIA_ValuePatternId)
-                    {
-                        return true;
-                    }
-
-                    return base.IsPatternSupported(patternId);
-                }
-
-                public override string Name
-                {
-                    get
-                    {
-                        string name = Owner.AccessibleName;
-                        if (name != null)
-                        {
-                            return name;
-                        }
-                        else
-                        {
-                            GridEntry selectedGridEntry = propertyGridView.SelectedGridEntry;
-                            if (selectedGridEntry != null)
-                            {
-                                return selectedGridEntry.AccessibilityObject.Name;
-                            }
-                        }
-
-                        return base.Name;
-                    }
-
-                    set
-                    {
-                        base.Name = value;
-                    }
-                }
-
-                #region IValueProvider
-
-                internal override bool IsReadOnly
-                {
-                    get
-                    {
-                        return !(propertyGridView.SelectedGridEntry is PropertyDescriptorGridEntry propertyDescriptorGridEntry) || propertyDescriptorGridEntry.IsPropertyReadOnly;
-                    }
-                }
-
-                #endregion
-
-                internal override void SetFocus()
-                {
-                    RaiseAutomationEvent(NativeMethods.UIA_AutomationFocusChangedEventId);
-
-                    base.SetFocus();
-                }
-            }
-        }
-
-        internal interface IMouseHookClient
-        {
-            // return true if the click is handled, false
-            // to pass it on
-            bool OnClickHooked();
-        }
-
-        internal class MouseHook
-        {
-            private readonly PropertyGridView gridView;
-            private readonly Control control;
-            private readonly IMouseHookClient client;
-
-            internal int thisProcessID = 0;
-            private GCHandle mouseHookRoot;
-            private IntPtr mouseHookHandle = IntPtr.Zero;
-            private bool hookDisable = false;
-
-            private bool processing;
-
-            public MouseHook(Control control, IMouseHookClient client, PropertyGridView gridView)
-            {
-                this.control = control;
-                this.gridView = gridView;
-                this.client = client;
-#if DEBUG
-                callingStack = Environment.StackTrace;
-#endif
-            }
-
-#if DEBUG
-            readonly string callingStack;
-            ~MouseHook()
-            {
-                Debug.Assert(mouseHookHandle == IntPtr.Zero, "Finalizing an active mouse hook.  This will crash the process.  Calling stack: " + callingStack);
-            }
-#endif
-
-            public bool DisableMouseHook
-            {
-
-                set
-                {
-                    hookDisable = value;
-                    if (value)
-                    {
-                        UnhookMouse();
-                    }
-                }
-            }
-
-            public virtual bool HookMouseDown
-            {
-                get
-                {
-                    GC.KeepAlive(this);
-                    return mouseHookHandle != IntPtr.Zero;
-                }
-                set
-                {
-                    if (value && !hookDisable)
-                    {
-                        HookMouse();
-                    }
-                    else
-                    {
-                        UnhookMouse();
-                    }
-                }
-            }
-
-            public void Dispose()
-            {
-                UnhookMouse();
-            }
-
-            /// <summary>
-            ///  Sets up the needed windows hooks to catch messages.
-            /// </summary>
-            private void HookMouse()
-            {
-                GC.KeepAlive(this);
-                // Locking 'this' here is ok since this is an internal class.
-                lock (this)
-                {
-                    if (mouseHookHandle != IntPtr.Zero)
-                    {
-                        return;
-                    }
-
-                    if (thisProcessID == 0)
-                    {
-                        SafeNativeMethods.GetWindowThreadProcessId(new HandleRef(control, control.Handle), out thisProcessID);
-                    }
-
-                    NativeMethods.HookProc hook = new NativeMethods.HookProc(new MouseHookObject(this).Callback);
-                    mouseHookRoot = GCHandle.Alloc(hook);
-
-                    mouseHookHandle = UnsafeNativeMethods.SetWindowsHookEx(NativeMethods.WH_MOUSE,
-                                                               hook,
-                                                               NativeMethods.NullHandleRef,
-                                                               SafeNativeMethods.GetCurrentThreadId());
-                    Debug.Assert(mouseHookHandle != IntPtr.Zero, "Failed to install mouse hook");
-                    Debug.WriteLineIf(CompModSwitches.DebugGridView.TraceVerbose, "DropDownHolder:HookMouse()");
-                }
-            }
-
-            /// <summary>
-            ///  HookProc used for catch mouse messages.
-            /// </summary>
-            private IntPtr MouseHookProc(int nCode, IntPtr wparam, IntPtr lparam)
-            {
-                GC.KeepAlive(this);
-                if (nCode == NativeMethods.HC_ACTION)
-                {
-                    NativeMethods.MOUSEHOOKSTRUCT mhs = Marshal.PtrToStructure<NativeMethods.MOUSEHOOKSTRUCT>(lparam);
-                    if (mhs != null)
-                    {
-                        switch (unchecked((int)(long)wparam))
-                        {
-                            case WindowMessages.WM_LBUTTONDOWN:
-                            case WindowMessages.WM_MBUTTONDOWN:
-                            case WindowMessages.WM_RBUTTONDOWN:
-                            case WindowMessages.WM_NCLBUTTONDOWN:
-                            case WindowMessages.WM_NCMBUTTONDOWN:
-                            case WindowMessages.WM_NCRBUTTONDOWN:
-                            case WindowMessages.WM_MOUSEACTIVATE:
-                                if (ProcessMouseDown(mhs.hWnd, mhs.pt.X, mhs.pt.Y))
-                                {
-                                    return (IntPtr)1;
-                                }
-                                break;
-                        }
-
-                    }
-                }
-
-                return UnsafeNativeMethods.CallNextHookEx(new HandleRef(this, mouseHookHandle), nCode, wparam, lparam);
-            }
-
-            /// <summary>
-            ///  Removes the windowshook that was installed.
-            /// </summary>
-            private void UnhookMouse()
-            {
-                GC.KeepAlive(this);
-                // Locking 'this' here is ok since this is an internal class.
-                lock (this)
-                {
-                    if (mouseHookHandle != IntPtr.Zero)
-                    {
-                        UnsafeNativeMethods.UnhookWindowsHookEx(new HandleRef(this, mouseHookHandle));
-                        mouseHookRoot.Free();
-                        mouseHookHandle = IntPtr.Zero;
-                        Debug.WriteLineIf(CompModSwitches.DebugGridView.TraceVerbose, "DropDownHolder:UnhookMouse()");
-                    }
-                }
-            }
-
-            /*
-           * Here is where we force validation on any clicks outside the
-           */
-            private bool ProcessMouseDown(IntPtr hWnd, int x, int y)
-            {
-
-                // if we put up the "invalid" message box, it appears this
-                // method is getting called re-entrantly when it shouldn't be.
-                // this prevents us from recursing.
-                //
-                if (processing)
-                {
-                    return false;
-                }
-
-                IntPtr hWndAtPoint = hWnd;
-                IntPtr handle = control.Handle;
-                Control ctrlAtPoint = Control.FromHandle(hWndAtPoint);
-
-                // if it's us or one of our children, just process as normal
-                //
-                if (hWndAtPoint != handle && !control.Contains(ctrlAtPoint))
-                {
-                    Debug.Assert(thisProcessID != 0, "Didn't get our process id!");
-
-                    // make sure the window is in our process
-                    SafeNativeMethods.GetWindowThreadProcessId(new HandleRef(null, hWndAtPoint), out int pid);
-
-                    // if this isn't our process, unhook the mouse.
-                    if (pid != thisProcessID)
-                    {
-                        HookMouseDown = false;
-                        return false;
-                    }
-
-                    bool needCommit = false;
-
-                    // if this a sibling control (e.g. the drop down or buttons), just forward the message and skip the commit
-                    needCommit = ctrlAtPoint == null ? true : !gridView.IsSiblingControl(control, ctrlAtPoint);
-
-                    try
-                    {
-                        processing = true;
-
-                        if (needCommit)
-                        {
-                            if (client.OnClickHooked())
-                            {
-                                return true; // there was an error, so eat the mouse
-                            }
-                            /* This breaks all sorts of stuff.  Need to find a better way to do this but we can't figure
-                               out the scenario this addressed.
-                            else {
-                                // Returning false lets the message go to its destination.  Only
-                                // return false if there is still a mouse button down.  That might not be the
-                                // case if committing the entry opened a modal dialog.
-                                //
-                                MouseButtons state = Control.MouseButtons;
-                                return (int)state == 0;
-                            }
-                            */
-                        }
-                    }
-                    finally
-                    {
-                        processing = false;
-                    }
-
-                    // cancel our hook at this point
-                    HookMouseDown = false;
-                    //gridView.UnfocusSelection();
-                }
-                return false;
-            }
-
-            /// <summary>
-            ///  Forwards messageHook calls to ToolTip.messageHookProc
-            /// </summary>
-            private class MouseHookObject
-            {
-                internal WeakReference reference;
-
-                public MouseHookObject(MouseHook parent)
-                {
-                    reference = new WeakReference(parent, false);
-                }
-
-                public virtual IntPtr Callback(int nCode, IntPtr wparam, IntPtr lparam)
-                {
-                    IntPtr ret = IntPtr.Zero;
-                    try
-                    {
-                        MouseHook control = (MouseHook)reference.Target;
-                        if (control != null)
-                        {
-                            ret = control.MouseHookProc(nCode, wparam, lparam);
-                        }
-                    }
-                    catch
-                    {
-                        // ignore
-                    }
-                    return ret;
-                }
-            }
-        }
-
-        /// <summary>
-        ///  The accessible object class for a PropertyGridView. The child accessible objects
-        ///  are accessible objects corresponding to the property grid entries.
-        /// </summary>
-        [ComVisible(true)]
-        internal class PropertyGridViewAccessibleObject : ControlAccessibleObject
-        {
-            private readonly PropertyGridView _owningPropertyGridView;
-            private readonly PropertyGrid _parentPropertyGrid;
-
-            /// <summary>
-            ///  Construct a PropertyGridViewAccessibleObject
-            /// </summary>
-            public PropertyGridViewAccessibleObject(PropertyGridView owner, PropertyGrid parentPropertyGrid) : base(owner)
-            {
-                _owningPropertyGridView = owner;
-                _parentPropertyGrid = parentPropertyGrid;
-            }
-
-            /// <summary>
-            ///  Return the child element at the specified point, if one exists,
-            ///  otherwise return this element if the point is on this element,
-            ///  otherwise return null.
-            /// </summary>
-            /// <param name="x">x coordinate of point to check</param>
-            /// <param name="y">y coordinate of point to check</param>
-            /// <returns>Return the child element at the specified point, if one exists,
-            ///  otherwise return this element if the point is on this element,
-            ///  otherwise return null.
-            /// </returns>
-            internal override UnsafeNativeMethods.IRawElementProviderFragment ElementProviderFromPoint(double x, double y)
-            {
-                return HitTest((int)x, (int)y);
-            }
-
-            /// <summary>
-            ///  Request to return the element in the specified direction.
-            /// </summary>
-            /// <param name="direction">Indicates the direction in which to navigate.</param>
-            /// <returns>Returns the element in the specified direction.</returns>
-            internal override UnsafeNativeMethods.IRawElementProviderFragment FragmentNavigate(UnsafeNativeMethods.NavigateDirection direction)
-            {
-                if (_parentPropertyGrid.AccessibilityObject is PropertyGridAccessibleObject propertyGridAccessibleObject)
-                {
-                    UnsafeNativeMethods.IRawElementProviderFragment navigationTarget = propertyGridAccessibleObject.ChildFragmentNavigate(this, direction);
-                    if (navigationTarget != null)
-                    {
-                        return navigationTarget;
-                    }
-                }
-
-                if (_owningPropertyGridView.OwnerGrid.SortedByCategories)
-                {
-                    switch (direction)
-                    {
-                        case UnsafeNativeMethods.NavigateDirection.FirstChild:
-                            return GetFirstCategory();
-                        case UnsafeNativeMethods.NavigateDirection.LastChild:
-                            return GetLastCategory();
-                    }
-                }
-                else
-                {
-                    switch (direction)
-                    {
-                        case UnsafeNativeMethods.NavigateDirection.FirstChild:
-                            return GetChild(0);
-                        case UnsafeNativeMethods.NavigateDirection.LastChild:
-                            int childCount = GetChildCount();
-                            if (childCount > 0)
-                            {
-                                return GetChild(childCount - 1);
-                            }
-
-                            return null;
-                    }
-                }
-
-                return base.FragmentNavigate(direction);
-            }
-
-            /// <summary>
-            ///  Return the element that is the root node of this fragment of UI.
-            /// </summary>
-            internal override UnsafeNativeMethods.IRawElementProviderFragmentRoot FragmentRoot
-            {
-                get
-                {
-                    return _owningPropertyGridView.OwnerGrid.AccessibilityObject;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the accessible object for the currently focused grid entry.
-            /// </summary>
-            /// <returns>The accessible object for the currently focused grid entry.</returns>
-            internal override UnsafeNativeMethods.IRawElementProviderFragment GetFocus()
-            {
-                return GetFocused();
-            }
-
-            /// <summary>
-            ///  Request value of specified property from an element.
-            /// </summary>
-            /// <param name="propertyId">Identifier indicating the property to return</param>
-            /// <returns>Returns a ValInfo indicating whether the element supports this property, or has no value for it.</returns>
-            internal override object GetPropertyValue(int propertyID)
-            {
-                if (propertyID == NativeMethods.UIA_ControlTypePropertyId)
-                {
-                    return NativeMethods.UIA_TableControlTypeId;
-                }
-                else if (propertyID == NativeMethods.UIA_NamePropertyId)
-                {
-                    return Name;
-                }
-
-                return base.GetPropertyValue(propertyID);
-            }
-
-            public override string Name
-            {
-                get
-                {
-                    string name = Owner.AccessibleName;
-                    if (name != null)
-                    {
-                        return name;
-                    }
-                    else
-                    {
-                        return SR.PropertyGridDefaultAccessibleName;
-                    }
-                }
-            }
-
-            public override AccessibleRole Role
-            {
-                get
-                {
-                    AccessibleRole role = Owner.AccessibleRole;
-                    if (role != AccessibleRole.Default)
-                    {
-                        return role;
-                    }
-                    else
-                    {
-                        return AccessibleRole.Table;
-                    }
-                }
-            }
-
-            public AccessibleObject Next(GridEntry current)
-            {
-                int row = ((PropertyGridView)Owner).GetRowFromGridEntry(current);
-                GridEntry nextEntry = ((PropertyGridView)Owner).GetGridEntryFromRow(++row);
-                if (nextEntry != null)
-                {
-                    return nextEntry.AccessibilityObject;
-                }
-                return null;
-            }
-
-            internal AccessibleObject GetCategory(int categoryIndex)
-            {
-                GridEntry[] targetEntries = new GridEntry[1];
-                GridEntryCollection topLevelGridEntries = _owningPropertyGridView.TopLevelGridEntries;
-                var topLevelGridEntriesCount = topLevelGridEntries.Count;
-                if (topLevelGridEntriesCount > 0)
-                {
-                    GridItem targetEntry = topLevelGridEntries[categoryIndex];
-                    if (targetEntry is CategoryGridEntry categoryGridEntry)
-                    {
-                        return categoryGridEntry.AccessibilityObject;
-                    }
-                }
-
-                return null;
-            }
-
-            internal AccessibleObject GetFirstCategory()
-            {
-                return GetCategory(0);
-            }
-
-            internal AccessibleObject GetLastCategory()
-            {
-                GridEntryCollection topLevelGridEntries = _owningPropertyGridView.TopLevelGridEntries;
-                var topLevelGridEntriesCount = topLevelGridEntries.Count;
-                return GetCategory(topLevelGridEntries.Count - 1);
-            }
-
-            /// <summary>
-            ///  Gets the previous grid entry accessibility object.
-            /// </summary>
-            /// <param name="currentGridEntry">The current grid entry.</param>
-            /// <param name="gridEntryCollection">The grid entry collection.</param>
-            /// <param name="currentGridEntryFound">Indicates whether the current grid entry is found.</param>
-            /// <returns>The previous grid entry.</returns>
-            internal AccessibleObject GetPreviousGridEntry(GridEntry currentGridEntry, GridEntryCollection gridEntryCollection, out bool currentGridEntryFound)
-            {
-                GridEntry previousGridEntry = null;
-                currentGridEntryFound = false;
-
-                foreach (GridEntry gridEntry in gridEntryCollection)
-                {
-                    if (currentGridEntry == gridEntry)
-                    {
-                        // Set to true to return the previous iterable element.
-                        currentGridEntryFound = true;
-                        if (previousGridEntry != null)
-                        {
-                            // In the current iteration return previous entry if the current entry == iterated grid entry.
-                            return previousGridEntry.AccessibilityObject;
-                        }
-                        else
-                        {
-                            return null;
-                        }
-                    }
-                    else
-                    {
-                        previousGridEntry = gridEntry;
-                        if (gridEntry.ChildCount > 0)
-                        {
-                            AccessibleObject foundChild = GetPreviousGridEntry(currentGridEntry, gridEntry.Children, out currentGridEntryFound);
-                            if (foundChild != null)
-                            {
-                                // Return some down-level child if found.
-                                return foundChild;
-                            }
-                            else if (currentGridEntryFound)
-                            {
-                                // If the passed current is found but there is no next near this current.
-                                return null;
-                            }
-                        }
-                    }
-                }
-
-                return null;
-            }
-
-            /// <summary>
-            ///  Gets the next grid entry.
-            /// </summary>
-            /// <param name="currentGridEntry">The current grid entry.</param>
-            /// <param name="gridEntryCollection">The grid entry collection.</param>
-            /// <param name="currentGridEntryFound">Indicates whether the current grid entry is found.</param>
-            /// <returns>The next grid entry.</returns>
-            internal AccessibleObject GetNextGridEntry(GridEntry currentGridEntry, GridEntryCollection gridEntryCollection, out bool currentGridEntryFound)
-            {
-                currentGridEntryFound = false;
-
-                foreach (GridEntry gridEntry in gridEntryCollection)
-                {
-                    if (currentGridEntryFound)
-                    {
-                        // Return the next entry via IEnumerable.Next() if previous entry == passed current.
-                        return gridEntry.AccessibilityObject;
-                    }
-
-                    if (currentGridEntry == gridEntry)
-                    {
-                        // Set to true to return the next iterable element. (see above)
-                        currentGridEntryFound = true;
-                    }
-                    else if (gridEntry.ChildCount > 0)
-                    {
-                        AccessibleObject foundChild = GetNextGridEntry(currentGridEntry, gridEntry.Children, out currentGridEntryFound);
-                        if (foundChild != null)
-                        {
-                            // Return some down-level child if found.
-                            return foundChild;
-                        }
-                        else if (currentGridEntryFound)
-                        {
-                            // If the passed current is found but there is no next near this current.
-                            return null;
-                        }
-                    }
-                }
-
-                return null;
-            }
-
-            /// <summary>
-            ///  Gets the first child property.
-            /// </summary>
-            /// <param name="current">The current grid entry.</param>
-            /// <returns>The first child property.</returns>
-            internal AccessibleObject GetFirstChildProperty(CategoryGridEntry current)
-            {
-                if (current.ChildCount > 0)
-                {
-                    GridEntryCollection subGridEntry = current.Children;
-                    if (subGridEntry != null && subGridEntry.Count > 0)
-                    {
-                        GridEntry[] targetEntries = new GridEntry[1];
-                        try
-                        {
-                            _owningPropertyGridView.GetGridEntriesFromOutline(subGridEntry, 0, 0, targetEntries);
-                        }
-                        catch (Exception ex)
-                        {
-                            Debug.Fail(ex.ToString());
-                        }
-
-                        return targetEntries[0].AccessibilityObject;
-                    }
-                }
-
-                return null;
-            }
-
-            /// <summary>
-            ///  Gets the last child property.
-            /// </summary>
-            /// <param name="current">The current grid entry.</param>
-            /// <returns>The last child property.</returns>
-            internal AccessibleObject GetLastChildProperty(CategoryGridEntry current)
-            {
-                if (current.ChildCount > 0)
-                {
-                    GridEntryCollection subGridEntry = current.Children;
-                    if (subGridEntry != null && subGridEntry.Count > 0)
-                    {
-                        GridEntry[] targetEntries = new GridEntry[1];
-                        try
-                        {
-                            _owningPropertyGridView.GetGridEntriesFromOutline(subGridEntry, 0, subGridEntry.Count - 1, targetEntries);
-                        }
-                        catch (Exception ex)
-                        {
-                            Debug.Fail(ex.ToString());
-                        }
-
-                        return targetEntries[0].AccessibilityObject;
-                    }
-                }
-
-                return null;
-            }
-
-            /// <summary>
-            ///  Gets the next category.
-            /// </summary>
-            /// <param name="current">The current grid entry.</param>
-            /// <returns>The next category.</returns>
-            internal AccessibleObject GetNextCategory(CategoryGridEntry current)
-            {
-                int row = _owningPropertyGridView.GetRowFromGridEntry(current);
-
-                GridEntry nextEntry;
-
-                do
-                {
-                    nextEntry = _owningPropertyGridView.GetGridEntryFromRow(++row);
-                    if (nextEntry is CategoryGridEntry)
-                    {
-                        return nextEntry.AccessibilityObject;
-                    }
-                }
-                while (nextEntry != null);
-
-                return null;
-            }
-
-            public AccessibleObject Previous(GridEntry current)
-            {
-                int row = ((PropertyGridView)Owner).GetRowFromGridEntry(current);
-                GridEntry prevEntry = ((PropertyGridView)Owner).GetGridEntryFromRow(--row);
-                if (prevEntry != null)
-                {
-                    return prevEntry.AccessibilityObject;
-                }
-                return null;
-            }
-
-            /// <summary>
-            ///  Gets the previous category.
-            /// </summary>
-            /// <param name="current">The current grid entry.</param>
-            /// <returns>The previous category.</returns>
-            internal AccessibleObject GetPreviousCategory(CategoryGridEntry current)
-            {
-                int row = _owningPropertyGridView.GetRowFromGridEntry(current);
-
-                GridEntry previousEntry;
-
-                do
-                {
-                    previousEntry = _owningPropertyGridView.GetGridEntryFromRow(--row);
-                    if (previousEntry is CategoryGridEntry)
-                    {
-                        return previousEntry.AccessibilityObject;
-                    }
-                }
-                while (previousEntry != null);
-
-                return null;
-            }
-
-            /// <summary>
-            ///  Get the accessible child at the given index.
-            ///  The accessible children of a PropertyGridView are accessible objects
-            ///  corresponding to the property grid entries.
-            /// </summary>
-            public override AccessibleObject GetChild(int index)
-            {
-
-                GridEntryCollection properties = ((PropertyGridView)Owner).AccessibilityGetGridEntries();
-                if (properties != null && index >= 0 && index < properties.Count)
-                {
-                    return properties.GetEntry(index).AccessibilityObject;
-                }
-                else
-                {
-                    return null;
-                }
-            }
-
-            /// <summary>
-            ///  Get the number of accessible children.
-            ///  The accessible children of a PropertyGridView are accessible objects
-            ///  corresponding to the property grid entries.
-            /// </summary>
-            public override int GetChildCount()
-            {
-                GridEntryCollection properties = ((PropertyGridView)Owner).AccessibilityGetGridEntries();
-
-                if (properties != null)
-                {
-                    return properties.Count;
-                }
-                else
-                {
-                    return 0;
-                }
-            }
-
-            /// <summary>
-            ///  Get the accessible object for the currently focused grid entry.
-            /// </summary>
-            public override AccessibleObject GetFocused()
-            {
-
-                GridEntry gridEntry = ((PropertyGridView)Owner).SelectedGridEntry;
-                if (gridEntry != null && gridEntry.Focus)
-                {
-                    return gridEntry.AccessibilityObject;
-                }
-                return null;
-            }
-
-            /// <summary>
-            ///  Get the accessible object for the currently selected grid entry.
-            /// </summary>
-            public override AccessibleObject GetSelected()
-            {
-                GridEntry gridEntry = ((PropertyGridView)Owner).SelectedGridEntry;
-                if (gridEntry != null)
-                {
-                    return gridEntry.AccessibilityObject;
-                }
-                return null;
-            }
-
-            /// <summary>
-            ///  Get the accessible child at the given screen location.
-            ///  The accessible children of a PropertyGridView are accessible objects
-            ///  corresponding to the property grid entries.
-            /// </summary>
-            public override AccessibleObject HitTest(int x, int y)
-            {
-                // Convert to client coordinates
-                var pt = new Point(x, y);
-                UnsafeNativeMethods.ScreenToClient(new HandleRef(Owner, Owner.Handle), ref pt);
-
-                // Find the grid entry at the given client coordinates
-                //
-                Point pos = ((PropertyGridView)Owner).FindPosition(pt.X, pt.Y);
-                if (pos != PropertyGridView.InvalidPosition)
-                {
-                    GridEntry gridEntry = ((PropertyGridView)Owner).GetGridEntryFromRow(pos.Y);
-                    if (gridEntry != null)
-                    {
-
-                        // Return the accessible object for this grid entry
-                        //
-                        return gridEntry.AccessibilityObject;
-                    }
-                }
-
-                // No grid entry at this point
-                //
-                return null;
-            }
-
-            /// <summary>
-            ///  Navigate to another object.
-            /// </summary>
-            public override AccessibleObject Navigate(AccessibleNavigation navdir)
-            {
-
-                if (GetChildCount() > 0)
-                {
-                    // We're only handling FirstChild and LastChild here
-                    switch (navdir)
-                    {
-                        case AccessibleNavigation.FirstChild:
-                            return GetChild(0);
-                        case AccessibleNavigation.LastChild:
-                            return GetChild(GetChildCount() - 1);
-                    }
-                }
-                return null;    // Perform default behavior
-            }
-        }
-
-        internal class GridPositionData
-        {
-            readonly ArrayList expandedState;
-            readonly GridEntryCollection selectedItemTree;
-            readonly int itemRow;
-            readonly int itemCount;
-
-            public GridPositionData(PropertyGridView gridView)
-            {
-                selectedItemTree = gridView.GetGridEntryHierarchy(gridView.selectedGridEntry);
-                expandedState = gridView.SaveHierarchyState(gridView.topLevelGridEntries);
-                itemRow = gridView.selectedRow;
-                itemCount = gridView.totalProps;
-            }
-
-            public GridEntry Restore(PropertyGridView gridView)
-            {
-                gridView.RestoreHierarchyState(expandedState);
-                GridEntry entry = gridView.FindEquivalentGridEntry(selectedItemTree);
-
-                if (entry != null)
-                {
-                    gridView.SelectGridEntry(entry, true);
-
-                    int delta = gridView.selectedRow - itemRow;
-                    if (delta != 0 && gridView.ScrollBar.Visible)
-                    {
-                        if (itemRow < gridView.visibleRows)
-                        {
-                            delta += gridView.GetScrollOffset();
-
-                            if (delta < 0)
-                            {
-                                delta = 0;
-                            }
-                            else if (delta > gridView.ScrollBar.Maximum)
-                            {
-                                delta = gridView.ScrollBar.Maximum - 1;
-                            }
-                            gridView.SetScrollOffset(delta);
-                        }
-
-                    }
-                }
-                return entry;
-            }
         }
     }
 }
